pyproject.toml
[project]
name = "password-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cryptography>=46.0.2",
    "mysql-connector-python==8.0.31",
    "python-dotenv==0.20.0",
]
